+++ crypto/src/base58/mod.rs
@@ -1,10 +1,29 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0
 
-mod encodable;
-mod error;
-pub mod version_bytes;
+#[macro_export]
+macro_rules! impl_bs58_json {
+    ($ty:ty, $ty_json:ty) => {
+        impl $ty {
+            pub fn from_base58(
+                input: impl AsRef<[u8]>,
+            ) -> Result<Self, mina_serialization_types::errors::Error> {
+                let t = <$ty_json>::from_base58(input)?;
+                Ok(t.into())
+            }
 
-pub use bs58::{decode, encode};
-pub use encodable::*;
-pub use error::*;
+            pub fn into_base58_string(
+                self,
+            ) -> Result<String, mina_serialization_types::errors::Error> {
+                let t: $ty_json = self.into();
+                t.to_base58_string()
+            }
+
+            pub fn to_base58_string(
+                &self,
+            ) -> Result<String, mina_serialization_types::errors::Error> {
+                self.clone().into_base58_string()
+            }
+        }
+    };
+}
